+++ .tests/log.test.js
@@ -10,7 +10,7 @@ describe('loggin.Log', () => {
 
     it(`construct formatter correctly`, () => {
         expect(() => {
-            new loggin.Log();
+            new loggin.Log('test');
         }).not.toThrow();
     });
 
@@ -23,15 +23,5 @@ describe('loggin.Log', () => {
                 }
             });
         }).not.toThrow();
-    });
-
-    it(`.format should work correctly`, () => {
-        let log = new loggin.Log();
-        expect(log.format()).toEqual();
-    });
-
-    it(`Log.fromObject should work correctly`, () => {
-        let log = loggin.Log.fromObject({  });
-        expect(log).toEqual();
     });
 });