+++ examples/readme/simple.js
@@ -1,0 +1,2 @@
+const loggin = require('../../src/node');
+loggin.debug('Check this log out!!', { foo: 'var' }, { channel: 'my-logger' });